fixed the unit test problem and changed files
